Make steering work from afar
Steering on reset didn't work when the two devices were afar, even
if the end device had already joined the network. Once the steering was
done the single messages came through just fine for way bigger
distances.

Looks like it was because the stack refused to initiate the steering if
LQI was below a threshold.

Remove the minimum threshold and bump esp-zigbee-lib to 1.6.1 which comes
with esp_zb_secur_network_min_join_lqi_set() making it possible.