Don't blink the led on the director node
We are not sure if the end device node is receiving the messages,
so the blinking to indicate the "connection" sometimes is unreliable.
Not blinking also means that the led always indicates the occupancy
status as soon as the device is powered on.

The current status is shown on the end device anyway.